rename details -> readmore
