+++ @navikt/core/react/src/index.ts
@@ -13,7 +13,7 @@ export * from "./modal";
 export * from "./pagination";
 export * from "./panel";
 export * from "./popover";
-export * from "./details";
+export * from "./read-more";
 export * from "./speech-bubble";
 export * from "./step-indicator";
 export * from "./stepper";
