+++ @navikt/core/react/src/read-more/ReadMore.tsx
