+++ @navikt/core/react/src/read-more/index.ts
